CheckedCross deleted, CheckedMark moved to assets\icons
