+++ src/components/ui/checkbox/Checkbox.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import { ComponentPropsWithoutRef, ReactNode } from 'react';
 
 import * as Checkbox from '@radix-ui/react-checkbox';
 
@@ -7,7 +7,9 @@ import s from './Checkbox.module.scss';
 type Props = {
   icon?: ReactNode;
   id?: string;
-} & Parameters<typeof Checkbox.Root>[0];
+} & ComponentPropsWithoutRef<typeof Checkbox.Root>;
+
+// type Props = ComponentPropsWithoutRef<typeof Checkbox.Root>;
 export const CheckboxRadix = (props: Props) => {
   const { checked, className, disabled, icon, id } = props;
